Test issue #3 (prior to any changes)
Before you decide whether or not to fix this issue, I'd like you to run the test suite with the two test cases that I've added in this patch. It will show the current behavior, with and without RaiseError being set, of calling a function that should receive an integer parameter where text has been passed instead.