Support retrying a function call in some cases
This can be achieved by passing (RaiseError => 0) as a second
parameter to CallFunction->new().  Note that CallFunction will still
die on some error conditions, so the caller should be prepared to deal
with that, too.

Also add support for retrying on serialization failures and
deadlocks into pg_proc_jsonrpc.psgi.

Documentation changes to follow.