Document changes in my previous commit
I chose not to elaborate on the new RaiseError parameter in
CallFunction.pm; adding it in without rewriting the documentation
seemed clumsy and the documentation already points to
pg_proc_jsonrpc.psgi for an example usage.  People can see it in
action there.

Also add a dependency on DBIx::Connector.

Bump to version 0.014.