Introduce a cache for proretset lookups
This can avoid one roundtrip to the server per function call.  On
my system, it resulted in roughly 40% performance improvement when
calling a relatively simple set-returning function.

The way we currently handle invalidations is to look at the SQLSTATEs
after trying to call the function, and if we see any suspicious
error codes, we invalidate the entry for that particular function.
This can result in some entries being falsely invalidated, but that
won't be a huge problem.  Also the caller needs to know that there's
a possibility of getting an error from postgres instead of CallFunction
when using the cache.

Also contains some whitespace fixes.