Avoid special pre- and post-processing for GetViewParams
With the "returns_json" changes this isn't necessary any more.  We
still need to have the mapping of GetViewParams->get_view() in place,
but that is fine.  Although that currently uses the full pre-/
post-processing infrastructure, I plan on getting rid of that infra
in favor of just a simple methodname->functionname mapping.  I'm
just not comfortable doing that until we have the entire API up and
running.